fileconnection: Expose file-contents differ error to user when using ckpt-open-files
When using the `--checkpoint-open-files` flag, on restart, the contents
of the saved version of a file are compared with its existing copy (if
any), and if the contents differ, DMTCP's policy is to throw an error to
the user. This is done to ensure that a user's file is not accidentally
overwritten. However, the existing implementation of the file comparison
function, `areFilesEqual()`, would assert out with a misleading error
message leaving the user confused when the file contents differed. This
patch forces the comparison function to return a comparison failed
(false) result to its caller to allow the caller to assert out with a
more meaningful error message.